+++ .dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/debug_bundle_linux-x64_assets.stamp
@@ -1,0 +1,1 @@
+{"inputs":["/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/.dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/app.dill","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/pubspec.yaml","/home/kevin/.pub-cache/hosted/pub.dev/cupertino_icons-1.0.6/assets/CupertinoIcons.ttf","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/material_fonts/MaterialIcons-Regular.otf","/home/kevin/snap/flutter/common/flutter/packages/flutter/lib/src/material/shaders/ink_sparkle.frag","/home/kevin/.pub-cache/hosted/pub.dev/async-2.11.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/boolean_selector-2.1.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/characters-1.3.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/clock-1.1.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/collection-1.18.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/cupertino_icons-1.0.6/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/fake_async-1.3.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/flutter_lints-3.0.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/leak_tracker-10.0.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/leak_tracker_flutter_testing-2.0.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/leak_tracker_testing-2.0.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/lints-3.0.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/matcher-0.12.16+1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/material_color_utilities-0.8.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/meta-1.11.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/path-1.9.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/source_span-1.10.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/stack_trace-1.11.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/stream_channel-2.1.2/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/string_scanner-1.2.0/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/term_glyph-1.2.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/test_api-0.6.1/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/vector_math-2.1.4/LICENSE","/home/kevin/.pub-cache/hosted/pub.dev/vm_service-13.0.0/LICENSE","/home/kevin/snap/flutter/common/flutter/bin/cache/pkg/sky_engine/LICENSE","/home/kevin/snap/flutter/common/flutter/packages/flutter/LICENSE"],"outputs":["/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/kernel_blob.bin","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/fonts/MaterialIcons-Regular.otf","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/shaders/ink_sparkle.frag","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/AssetManifest.json","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/AssetManifest.bin","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/FontManifest.json","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/NOTICES.Z","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/build/flutter_assets/version.json"]}
