+++ .dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/native_assets.stamp
@@ -1,0 +1,1 @@
+{"inputs":["/home/kevin/snap/flutter/common/flutter/packages/flutter_tools/lib/src/build_system/targets/native_assets.dart","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/.dart_tool/package_config_subset"],"outputs":["/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/.dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/native_assets.yaml","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/.dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/native_assets.yaml"]}
