+++ .dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/native_assets.yaml
@@ -1,0 +1,5 @@
+format-version:
+  - 1
+  - 0
+  - 0
+native-assets: {}
