+++ .dart_tool/flutter_build/e290cd833db85fff90ee46becd2946f3/unpack_linux.stamp
@@ -1,0 +1,1 @@
+{"inputs":["/home/kevin/snap/flutter/common/flutter/packages/flutter_tools/lib/src/build_system/targets/linux.dart","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/libflutter_linux_gtk.so","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/icudtl.dat","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_plugin_registrar.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_dart_project.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_binary_messenger.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_method_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_json_method_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_event_channel.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_string_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_value.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_message_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_basic_message_channel.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_texture_registrar.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_engine.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_binary_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/flutter_linux.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_texture_gl.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_method_channel.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_standard_message_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_pixel_buffer_texture.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_texture.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_view.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_method_response.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_json_message_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_standard_method_codec.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_method_call.h","/home/kevin/snap/flutter/common/flutter/bin/cache/artifacts/engine/linux-x64/flutter_linux/fl_plugin_registry.h"],"outputs":["/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/libflutter_linux_gtk.so","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/icudtl.dat","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_plugin_registrar.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_dart_project.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_binary_messenger.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_method_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_json_method_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_event_channel.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_string_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_value.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_message_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_basic_message_channel.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_texture_registrar.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_engine.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_binary_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/flutter_linux.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_texture_gl.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_method_channel.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_standard_message_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_pixel_buffer_texture.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_texture.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_view.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_method_response.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_json_message_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_standard_method_codec.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_method_call.h","/home/kevin/Documents/3º ano/2º Semestre/ES/ES-2024-LEI-Quarta-Grupo-E/linux/flutter/ephemeral/flutter_linux/fl_plugin_registry.h"]}
